refactor: only try to purge the cache when invalidation is configured

+++ packages/cloudflare/src/api/overrides/tag-cache/d1-next-tag-cache.ts
@@ -2,7 +2,7 @@ import { error } from "@opennextjs/aws/adapters/logger.js";
 import type { NextModeTagCache } from "@opennextjs/aws/types/overrides.js";
 
 import { getCloudflareContext } from "../../cloudflare-context.js";
-import { debugCache, FALLBACK_BUILD_ID, purgeCacheByTags } from "../internal.js";
+import { debugCache, FALLBACK_BUILD_ID, isPurgeCacheEnabled, purgeCacheByTags } from "../internal.js";
 
 export const NAME = "d1-next-mode-tag-cache";
 
@@ -67,7 +67,9 @@ export class D1NextModeTagCache implements NextModeTagCache {
 		);
 
 		// TODO: See https://github.com/opennextjs/opennextjs-aws/issues/986
-		await purgeCacheByTags(tags);
+		if (isPurgeCacheEnabled()) {
+			await purgeCacheByTags(tags);
+		}
 	}
 
 	private getConfig() {
